pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "aliccp-mtl-ranker"
version = "0.1.0"
description = "Canonical preprocessing pipeline for Ali-CCP dataset."
requires-python = ">=3.9"
dependencies = [
    "pandas>=2.0.0",
    "pyarrow>=12.0.0",
    "pyyaml>=6.0.0",
    "duckdb>=1.1.1",
    "matplotlib>=3.9.3",
    "seaborn>=0.13.2",
    "tqdm>=4.66.5",
]

[tool.pytest.ini_options]
pythonpath = ["."]
